test/requirements.txt
beautifulsoup4==4.4.1
cryptography
requests==2.8.1
sanction==0.4.1
pytest==2.9.1
pytest-xdist==1.15.0
PyJWT==1.6.4
